modified some components
