+++ app/layout.tsx
@@ -1,15 +1,21 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { Oswald } from "next/font/google";
+import { Poppins } from "next/font/google";
 import "destyle.css";
 import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
-const OswaldFont = Oswald({
+const oswald = Oswald({
 	weight: "400",
 	subsets: ["latin"],
 	variable: "--font-Oswald",
 });
+const poppins = Poppins({
+	weight: "600",
+	subsets: ["latin"],
+	variable: "--font-Poppins",
+})
 
 export const metadata: Metadata = {
 	title: "Gemi21",
@@ -21,7 +27,7 @@ export default function RootLayout({
 }: Readonly<{ children: React.ReactNode }>) {
 	return (
 		<html lang="en">
-			<body className={`${inter.className} ${OswaldFont.variable}`}>
+			<body className={`${inter.className} ${oswald.variable} ${poppins.variable}`}>
 				{children}
 			</body>
 		</html>
