+++ app/types/Types.d.ts
@@ -1,1 +1,1 @@
-declare module 'tailwindcss/lib/util/flattenColorPalette';
+declare module "tailwindcss/lib/util/flattenColorPalette";
