+++ package.json
@@ -12,6 +12,7 @@
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-navigation-menu": "^1.2.3",
     "@radix-ui/react-slot": "^1.1.0",
+    "@tabler/icons-react": "^3.26.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "destyle.css": "^4.0.1",
@@ -34,7 +35,6 @@
     "eslint": "^8",
     "eslint-config-next": "14.2.5",
     "postcss": "^8",
-    "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
